// src/app/enums/enums.ts
// src/app/enums/enums.ts

export enum Arbeitsverhaeltnis {
  Befristet   = "Befristet",
  Unbefristet = "Unbefristet"
}

export enum Bereich {
  Direkt   = "Direkt",
  Indirekt = "Indirekt"
}

export enum Mengenabhaengigkeit {
  Mengenabhaengig   = "Mengenabhaengig",
  Mengenunabhaengig = "Mengenunabhaengig"
}

export enum Austrittsart {
  KeineAngabe = 'KeineAngabe',
  ArbeitnehmerKündigung = 'ArbeitnehmerKündigung',
  ArbeitgeberKündigung = 'ArbeitgeberKündigung',
  Altersteilzeit = 'Altersteilzeit',
  Sonstiges = 'Sonstiges'
}